+++ src/opcode.h
@@ -74,6 +74,7 @@ namespace arsh {
   OP(DUP, 0, 1)                                                                                    \
   OP(DUP2, 0, 2)                                                                                   \
   OP(SWAP, 0, 0)                                                                                   \
+  OP(STORE_BY_OFFSET, 1, -1)                                                                       \
   OP(CONCAT, 0, -1)                                                                                \
   OP(APPEND, 0, -1)                                                                                \
   OP(APPEND_ARRAY, 0, -1)                                                                          \